Made Changes to Augmented Reality Example
Removed the second object that Emma had been using to try to make an
arrow because it was behaving strangely.  Changed texture of object to
solid color.  Changed shape of object to a frustum like prism.  Added a
second directional light to the virtual scene (since the lack of
virtual light in some direction was making viewing the model from some
directions weird) but Tango died before I had a chance to test this
more thoroughly.  Was working on trying to control the rotation of the
object (rotation of the object and how it sits on planes or maybe
doesn’t sit on planes is something we will have to think about.)

Also a couple changes made to MAP by accident I think but since that
project is obsolete it shouldn’t matter…