Added Cubes and a line to demonstrate adding AR objects
Found the place in the code to add objects after clicking on the
location to navigate to.  Can dynamically add multiple objects to
specific places in the real world.  ToDo: need to make sure 2d map, 3d
adf, and locations all work, and most importantly need to make sure
that we can reliably give the renderer the correct locations to create
the AR objects.